+++ trunk/app/CECRunningChart.Web/Models/Runningchart/ChartPumpstationItemModel.cs
@@ -10,7 +10,7 @@ namespace CECRunningChart.Web.Models.Runningchart
     {
         //public int PumpstationId { get; set; }
         //public string PumpStationName { get; set; }
-        public int PumpAmount { get; set; }
+        public decimal PumpAmount { get; set; }
         public int SelectedPumpstationId { get; set; }
         public bool IsRemoving { get; set; }
     }
